Move 'introspectionQuery.js' to 'getIntrospectionQuery.js' (#2125)
