+++ src/type/__tests__/introspection-test.js
@@ -8,7 +8,7 @@ import invariant from '../../jsutils/invariant';
 import { missingFieldArgMessage } from '../../validation/rules/ProvidedRequiredArguments';
 
 import { graphqlSync } from '../../graphql';
-import { getIntrospectionQuery } from '../../utilities/introspectionQuery';
+import { getIntrospectionQuery } from '../../utilities/getIntrospectionQuery';
 
 import { GraphQLSchema } from '../schema';
 import { GraphQLString } from '../scalars';
